Person row working (#8)
* Copied movie html files and java classes, and renamed them to fit person

* Updated person_reg.html

* Changed person_reg.html select menu values integer to string role

* Changed m_id to p_id in person_reg.html

* Updated person_reg.html to match Joseph's format for movie_reg.html

* Matched person_upd.html to person_reg.html

* Added missing <p> tag

* Matched person_del.html with person_reg.html

* Update person_reg.html

Now includes the necessary variables for a person to not only be added to the person table but the role table of the person (i.e. adds actor John Doe to the person and actor tables as an entry).

* Update person_reg.java

Incomplete, but can successfully add to a person to the person table its respective role table should there be no errors thrown. Still needs to move some code to a separate method. Will update this again later.

* Update person_reg.java

Adds a person entry, then a role table entry determined by the role_name value entered. If it fails to enter into the role table, the person entry is also deleted.

* Update person_reg.java

Forgot to change the username and password.

* Update person_reg.java

Should be a working file. Further changes would be for sake of optimization/cleanup.

* Update person_del.java

Currently a mess of if-else, but works. Will concise this down later.

* Update person_upd.java

Note - for updating a role entry, we can only update the position. If one wants to change the movie ID for an entry, that will not work with this. You'd instead have to delete the entry, and recreate a new one.

* Update person_del.html

Separated deleting a person and a role entry.

* Update person_reg.html

Made role_ID and p_ID the same,

* Update person_upd.html

Separated updating a role entry and a person entry.

* Update personJdbc.java

Works now.

* Copied movie html files and java classes, and renamed them to fit person

* Updated person_reg.html

* Changed person_reg.html select menu values integer to string role

* Changed m_id to p_id in person_reg.html

* Updated person_reg.html to match Joseph's format for movie_reg.html

* Matched person_upd.html to person_reg.html

* Added missing <p> tag

* Matched person_del.html with person_reg.html

* Update person_reg.html

Now includes the necessary variables for a person to not only be added to the person table but the role table of the person (i.e. adds actor John Doe to the person and actor tables as an entry).

* Update person_reg.java

Incomplete, but can successfully add to a person to the person table its respective role table should there be no errors thrown. Still needs to move some code to a separate method. Will update this again later.

* Update person_reg.java

Adds a person entry, then a role table entry determined by the role_name value entered. If it fails to enter into the role table, the person entry is also deleted.

* Update person_reg.java

Forgot to change the username and password.

* Update person_reg.java

Should be a working file. Further changes would be for sake of optimization/cleanup.

* Update person_del.java

Currently a mess of if-else, but works. Will concise this down later.

* Update person_upd.java

Note - for updating a role entry, we can only update the position. If one wants to change the movie ID for an entry, that will not work with this. You'd instead have to delete the entry, and recreate a new one.

* Update person_del.html

Separated deleting a person and a role entry.

* Update person_reg.html

Made role_ID and p_ID the same,

* Update person_upd.html

Separated updating a role entry and a person entry.

* Update personJdbc.java

Works now.

* Changed username and password to 'c##project' and 'project' respectfully

---------

Co-authored-by: JCSI2022 <113455662+LF92R@users.noreply.github.com>